Define a default content part display driver and refactor underlying abstract class to simplify and reduce duplicated code
Check that the contentPart argument is of type ContentPart (not any subclasses of ContentPart). If there is no ContentPart subclass representation, then the part must have been defined dynamically via the admin dashboard, which are the parts that we want

Then check to see if the part is an implicit part. If so ignore, as it doesn't make sense to provide a part template as this would constrain the fields, which should be able to be rendered anywhere for the entire content item.

The rest of the logic is the same as the underlying abstract class

Finally register the DefaultContentPartDisplayDriver with the service provider