Make sure fields for explicit parts get added to the part shape, rather than the zone / holding shape
When resolving parent shape in ShapeResult ApplyImplementationAsync, check if context arg implements IFieldDisplayManagementContext, indicating that this ShapeResult object represents a field shape. Try and get the PartName that the field belongs to. Then get the shape for that Part, which will already exist on the zoneholding shape (the parentShape that has been resolved so far), given that Part shapes are handled before Field shapes by the content item display coordinator.

To facilitate checking the current ShapeResult object represents a field shape, a new interface IFieldDisplayManagementContext is defined in OrchardCore.DisplayManagement, and implemented by BuildFieldDisplayContext in Orchardcore.ContentManagement. This is to avoid OrchardCore.DisplayManagement needing to take a dependency on OrchardCore.ContentManagement (which itself depends on OrchardCore.DisplayManagement) - this would be a circular dependency.